Added SwiftData, but commented out some features because of the problems with beta - Notification Time - Custom Color - Query sorting - #Previews

+++ Pet Reminder/Helper Classes/Enums/NotificationSelection.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-enum NotificationSelection: Int64 {
+enum NotificationSelection: Int {
     case both
     case morning
     case evening
